ComputeStep as interface
