+++ algo/src/main/java/org/neo4j/graphalgo/impl/pagerank/NonWeightedComputeStepFactory.java
@@ -5,8 +5,8 @@ import org.neo4j.graphalgo.api.RelationshipIterator;
 import org.neo4j.graphalgo.api.RelationshipWeights;
 
 public class NonWeightedComputeStepFactory implements  ComputeStepFactory {
-    public ComputeStep createComputeStep(double dampingFactor, int[] sourceNodeIds, RelationshipIterator relationshipIterator, Degrees degrees, RelationshipWeights relationshipWeights, int partitionCount, int start) {
-        return new ComputeStep(
+    public NonWeightedComputeStep createComputeStep(double dampingFactor, int[] sourceNodeIds, RelationshipIterator relationshipIterator, Degrees degrees, RelationshipWeights relationshipWeights, int partitionCount, int start) {
+        return new NonWeightedComputeStep(
                 dampingFactor,
                 sourceNodeIds,
                 relationshipIterator,
